Fix edge case for zero-length decoration in empty leaf.
This change ensures that a zero-length decoration is never treated as encompassing the entire leaf,
handling an edge case where an empty decoration inside an empty leaf was being treated
as encompassing the entire leaf and thus preventing the "before" and "after" leaves
we would otherwise see (when a zero-length decoration is applied to any non-empty leaf).

This fixes issues encountered in the `slate-page-decorations` library for the edge case
of an empty paragraph node being located on the edge of a page boundary.