Remove unnecessary upstream GitHub Actions for this fork.
This fork won't use the same release process as the upstream,
and also won't need to use the same "magic comments" automation for PRs.

We remove these workflows on our fork both because they are unnecessary
and because they represent a security risk - they make use of secrets
by name, and we wouldn't want these to accidentally match any org
secrets that we currently have set up in our org, since that would
allow them to take actions using our NPM credentials (or similar).

For when we eventually do a PR to the upstream repo, we want to
avoid including this commit in the PR - we want to only include
the commits with the code fixes we have worked on.